# Backend/docs/openapi-final.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIGGAB API",
    "version": "2.1.0",
    "description": "Sistema Gestor de Ganado Bovino (SIGGAB) - API REST.\n\nTodos los endpoints (excepto login y refresh) requieren Bearer token.\nRespuesta estandar: { success, data, message, errors }.",
    "contact": {
      "name": "SIGGAB Dev Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido en POST /auth/login"
      }
    },
    "schemas": {
...
      "CreateUsuarioRequest": {
        "type": "object",
        "required": [
          "nombreCompleto",
          "username",
          "password",
          "idRol"
        ],
        "properties": {
          "nombreCompleto": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100
          },
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "idRol": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
      "TipoInsumo": {
        "type": "object",
        "properties": {
          "idTipoInsumo": {
            "type": "integer"
          },
          "nombreTipo": {
            "type": "string"
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "activo": {
            "type": "boolean"
          }
        }
      },
      "Insumo": {
        "type": "object",
        "properties": {
          "idInsumo": {
            "type": "integer"
          },
          "nombreInsumo": {
            "type": "string"
          },
          "idTipoInsumo": {
            "type": "integer"
          },
          "unidadMedida": {
            "type": "string"
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "stockActual": {
            "type": "number"
          },
          "activo": {
            "type": "boolean"
          },
          "tipoInsumo": {
            "$ref": "#/components/schemas/TipoInsumo"
          }
        }
      },
      "DetalleSolicitud": {
        "type": "object",
        "properties": {
          "idInsumo": {
            "type": "integer"
          },
          "cantidad": {
            "type": "number"
          },
          "precioEstimado": {
            "type": "number"
          },
          "subtotalEstimado": {
            "type": "number"
          },
          "insumo": {
            "$ref": "#/components/schemas/Insumo"
          }
        }
      },
      "SolicitudCompra": {
        "type": "object",
        "properties": {
          "idSolicitud": {
            "type": "integer"
          },
          "fechaSolicitud": {
            "type": "string",
            "format": "date"
          },
          "estadoSolicitud": {
            "type": "string",
            "enum": [
              "PENDIENTE",
              "APROBADA",
              "RECHAZADA"
            ]
          },
          "observaciones": {
            "type": "string",
            "nullable": true
          },
          "detalles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleSolicitud"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/solicitudes-compra": {
      "get": {
        "tags": [
          "Compras"
        ],
        "summary": "Listar solicitudes de compra",
        "parameters": [
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "APROBADA",
                "RECHAZADA"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de solicitudes con sus detalles"
          }
        }
      },
      "post": {
        "tags": [
          "Compras"
        ],
        "summary": "Crear solicitud de compra (solo Almacén — RN-13)",
        "description": "**RN-13:** Solo el rol **Almacén** puede crear solicitudes.\nIncluye detalles con insumos, cantidades y precios estimados.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudCompra"
              },
              "example": {
                "fechaSolicitud": "2024-07-01",
                "observaciones": "Reposición mensual de medicamentos",
                "detalles": [
                  {
                    "idInsumo": 1,
                    "cantidad": 50,
                    "precioEstimado": 12.5
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Solicitud creada con detalles"
          },
          "403": {
            "description": "Solo Almacén puede crear solicitudes (RN-13)"
          }
        }
      }
    },
    "/solicitudes-compra/{id}": {
      "get": {
        "tags": [
          "Compras"
        ],
        "summary": "Obtener solicitud por ID con detalles",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitud con detalles e historial de compras"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      }
    },
    "/solicitudes-compra/{id}/aprobar": {
      "patch": {
        "tags": [
          "Compras"
        ],
        "summary": "Aprobar o rechazar solicitud (solo Admin — RN-14)",
        "description": "**RN-14:** Solo el **Administrador** puede aprobar/rechazar.\n**RN-15:** Una solicitud RECHAZADA no puede reabrirse.\n**RN-17:** Una solicitud con compra ya generada no puede cambiarse.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "estadoSolicitud"
                ],
                "properties": {
                  "estadoSolicitud": {
                    "type": "string",
                    "enum": [
                      "APROBADA",
                      "RECHAZADA"
                    ]
                  },
                  "observaciones": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitud procesada"
          },
          "400": {
            "description": "Ya fue rechazada o ya tiene compra (RN-15, RN-17)"
          },
          "403": {
            "description": "Solo Administrador (RN-14)"
          }
        }
      }
    },
...
    "/usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar usuarios (solo Admin)",
        "responses": {
          "200": {
            "description": "Lista de usuarios"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Crear usuario (solo Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsuarioRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado"
          },
          "409": {
            "description": "Username ya en uso"
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtener usuario por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del usuario"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Actualizar datos del usuario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombreCompleto": {
                    "type": "string"
                  },
                  "idRol": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado"
          }
        }
      }
    },
    "/usuarios/{id}/estado": {
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Activar o desactivar usuario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "activo"
                ],
                "properties": {
                  "activo": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          }
        }
      }
    }
  },
...
}